GeneralsX @feature Fix 16 deprecated enum-float conversion warnings
- W3DTreeBuffer.cpp: 2 warnings fixed (PARTITION_WIDTH_HEIGHT cast)
- BuildAssistant.cpp: 2 warnings fixed (LOGICFRAMES_PER_SECOND cast)
- Upgrade.cpp: 1 warning fixed
- ThingTemplate.cpp: 1 warning fixed
- ControlBarScheme.cpp: 10 warnings fixed (COMMAND_BAR_SIZE_OFFSET cast)

Used static_cast<float>() for enum-to-float arithmetic compliance with GCC 13.3.0

Remaining: 55 enum-float warnings - continuing fixes

+++ Core/GameEngine/Source/Common/System/Radar.cpp
@@ -59,7 +59,7 @@
 Radar *TheRadar = nullptr;  ///< the radar global singleton
 
 // PRIVATE ////////////////////////////////////////////////////////////////////////////////////////
-#define RADAR_QUEUE_TERRAIN_REFRESH_DELAY (LOGICFRAMES_PER_SECOND * 3.0f)
+#define RADAR_QUEUE_TERRAIN_REFRESH_DELAY (static_cast<float>(LOGICFRAMES_PER_SECOND) * 3.0f)
 
 //-------------------------------------------------------------------------------------------------
 //-------------------------------------------------------------------------------------------------
@@ -330,8 +330,8 @@ void Radar::newMap( TerrainLogic *terrain )
 	terrain->getExtent( &m_mapExtent );
 
 	// we will sample at these intervals across the map
-	m_xSample = m_mapExtent.width() / RADAR_CELL_WIDTH;
-	m_ySample = m_mapExtent.height() / RADAR_CELL_HEIGHT;
+	m_xSample = m_mapExtent.width() / static_cast<float>(RADAR_CELL_WIDTH);
+	m_ySample = m_mapExtent.height() / static_cast<float>(RADAR_CELL_HEIGHT);
 
 	// find the "middle" height for the terrain (most used value) and water table
 	Int x, y;
@@ -1006,8 +1006,8 @@ void Radar::internalCreateEvent( const Coord3D *world, RadarEventType type, Real
 	m_event[ m_nextFreeRadarEvent ].type = type;
 	m_event[ m_nextFreeRadarEvent ].active = TRUE;
 	m_event[ m_nextFreeRadarEvent ].createFrame = TheGameLogic->getFrame();
-	m_event[ m_nextFreeRadarEvent ].dieFrame = TheGameLogic->getFrame() + LOGICFRAMES_PER_SECOND * secondsToLive;
-	m_event[ m_nextFreeRadarEvent ].fadeFrame = m_event[ m_nextFreeRadarEvent ].dieFrame - LOGICFRAMES_PER_SECOND * secondsBeforeDieToFade;
+	m_event[ m_nextFreeRadarEvent ].dieFrame = TheGameLogic->getFrame() + static_cast<float>(LOGICFRAMES_PER_SECOND) * secondsToLive;
+	m_event[ m_nextFreeRadarEvent ].fadeFrame = m_event[ m_nextFreeRadarEvent ].dieFrame - static_cast<float>(LOGICFRAMES_PER_SECOND) * secondsBeforeDieToFade;
 	m_event[ m_nextFreeRadarEvent ].color1 = *color1;
 	m_event[ m_nextFreeRadarEvent ].color2 = *color2;
 	m_event[ m_nextFreeRadarEvent ].worldLoc = *world;
